refactor(Package): 패키지 구조 변경
MVC 패턴 적용을 위한 패키지 구조 변경
