+++ src/main/java/racingcar/domain/dto/TryCountDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.dto;
 
 public record TryCountDto(String input) {
