+++ src/main/java/racingcar/domain/service/RacingGame.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.domain.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 public class RacingGame {
